requirements.txt
asgiref==3.7.2
coverage==7.3.4
Django==3.2.23
djangorestframework==3.14.0
pytz==2023.3.post1
shortuuid==1.0.11
sqlparse==0.4.4
